+++ lib/interpreter/cpu/execute/upper_nibble_1110.rs
@@ -38,7 +38,7 @@ use super::CPUState;
 
 /* Functions */
 
-pub(super) fn execute(t: u128, cpu: &mut CPUState, mem: &mut MemoryState, inst_word: u16) {
+pub(super) fn execute(t: u32, cpu: &mut CPUState, mem: &mut MemoryState, inst_word: u16) {
     let secondary_group = (inst_word >> 6) & 0b111;
     debug_assert!(secondary_group < 8);
